+++ autoupdate/lib/tools/check_for_updates.dart
@@ -1,8 +1,0 @@
-import 'pixelarticons_version.dart';
-
-/// Returns [true] if there's a new [pixelarticons] update available (new unreleased icon pack)
-Future<bool> hasUpdateAvailable() async {
-  final String current = await getCurrentPixelarticonsVersion();
-  final String latest = await getLatestPixelarticonsVersion();
-  return current != latest;
-}